bake.py: add warning if building on unsupported os
Print a warning if building on unsupported os. The warning triggers a
prompt asking to continue or abort build. Use '-s' (--skip-os-check) option to
automatically continue build.

To add a supported OS see core/setup directory. Files can also be added in
a setup folder in local meta layer. To unsupport an OS add a
setup file with only the distro-release line (keep only the first line).

It would be handy to have a feature to disable all other setup folders
but the one in local layer. But so far this has not been implemented.